Split worker code into separate go file
